refactor: users 라우팅 / 컴포넌트 디렉토리 구조 수정 반영
