+++ src/app/portal/users/members/[id]/page.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'next/navigation';
 import { Button } from '@/components/ui/button';
-import { ArrowLeft, Edit } from 'lucide-react';
+import { ArrowLeft } from 'lucide-react';
 import Link from 'next/link';
 import {
   AdminUserApi,
@@ -11,8 +11,7 @@ import {
   UserDetail as UserDetailType,
   UserProjectItem,
 } from '@/generated-api';
-import AdminUserDetail from '@/components/system/users/user-detail';
-import UserEditModal from '@/components/system/users/user-edit-modal';
+import UserDetail from '@/components/portal/users/members/user-detail';
 import { getApiConfig } from '@/lib/config';
 
 const adminUserApi = new AdminUserApi(getApiConfig());
@@ -24,7 +23,6 @@ export default function UserDetailPage() {
   const userId = params.id as string;
   const [user, setUser] = useState<UserDetailType>();
   const [userProjects, setUserProjects] = useState<UserProjectItem[]>([]);
-  const [editModalOpen, setEditModalOpen] = useState(false);
 
   useEffect(() => {
     const fetchUser = async () => {
@@ -62,35 +60,15 @@ export default function UserDetailPage() {
       <div className="flex items-center justify-between">
         <div className="flex items-center gap-4">
           <Button variant="ghost" size="sm" asChild>
-            <Link href="/system/users">
+            <Link href="/portal/users/members">
               <ArrowLeft className="mr-2 h-4 w-4" />
               목록으로
             </Link>
           </Button>
         </div>
-
-        {/* 정보 수정 버튼 */}
-        <Button
-          onClick={() => {
-            setUser(user);
-            setEditModalOpen(true);
-          }}
-        >
-          <Edit className="mr-2 h-4 w-4" />
-          정보 수정
-        </Button>
       </div>
 
-      <AdminUserDetail user={user} projects={userProjects} />
-
-      <UserEditModal
-        user={user}
-        open={editModalOpen}
-        onOpenChange={setEditModalOpen}
-        onUserUpdate={(updatedUser) => {
-          setUser(updatedUser); // 상세 정보 갱신
-        }}
-      />
+      <UserDetail user={user} projects={userProjects} />
     </div>
   );
 }
