+++ src/components/portal/researches/projects/timeline/timeline-form-modal.tsx
@@ -35,7 +35,7 @@ import { FileSummary, TimelineSummary } from '@/generated-api/models';
 import { useAuthStore } from '@/store/auth-store';
 import { toast } from 'sonner';
 import { uploadFileWithPresignedUrl } from '@/lib/upload';
-import { FileItem } from '@/components/portal/researches/projects/file-item';
+import { FileItem } from '@/components/common/file-item';
 import TimePicker from '@/components/common/time-picker';
 import { getApiConfig } from '@/lib/config';
